Material 편집기 구현 중 남은 작업: Save, Load, 인게임 Material 수정 기능

+++ Project/Engine/Com_Animator3D.cpp
@@ -194,7 +194,7 @@ namespace mh
 			{
 				if (_animName == clips[i].strAnimName)
 				{
-					m_iCurClip = i;
+					m_iCurClip = (int)i;
 
 					//m_iFrameCount = 0;
 					//m_dCurTime = 0.f;
